// tab-frontend/src/app/modules/permissions/models/functional-block-enum.ts
export enum FUNCTIONALBLOCK {
    USERS = "USERS",
    ROLES = "ROLES",
    FUNCTIONALITY = "FUNCTIONALITY",
    ACCESSTYPES = "ACCESSTYPES",
    PERMISSIONS = "PERMISSIONS",
    BLOGS = "BLOGS",
    PROFILEPHOTO = "PROFILEPHOTO",
    COMPANIES = "COMPANIES",
    UNIVERSITIES = "UNIVERSITIES",
    SKILLS = "SKILLS",
    CERTCENTERS = "CERTCENTERS",
    MOTIVATIONFACTORS = "MOTIVATIONFACTORS",
    BENEFITS = "BENEFITS",
    CURRENCIES = "CURRENCIES",
    COUNTRIES = "COUNTRIES",
    CITIES = "CITIES",
    TALENTPROFILES = "TALENTPROFILES",
    TALENTACQUISITIONPROFILE = "TALENTACQUISITIONPROFILE",
    BRIEFS = "BRIEFS",
    MEETINGS = "MEETINGS",
    MEETINGINVITATIONS = "MEETINGINVITATIONS",
    MEETINGTEMPLATES = "MEETINGTEMPLATES",
    SCHEDULE = "SCHEDULE",
    SCHEDULEDEFAULTSETTINGS = "SCHEDULEDEFAULTSETTINGS",
    POSITIONS = "POSITIONS",
    SOCIALMEDIA = "SOCIALMEDIA",
    VISITORS = "VISITORS",
    JOBPLATFORMS = "JOBPLATFORMS",
    INTERVIEWS = "INTERVIEWS",
    CV = "CV",
    CVPDF = "CVPDF",
    VISITCARDSPRF = "VISITCARDSPRF",
    COVERLETTERS = "COVERLETTERS",
    AUTH = "AUTH",
    VERIFICATION = "VERIFICATION",
    BLOGIMAGES = "BLOGIMAGES",
    USERPROFILES = "USERPROFILES",
    LEAD = "LEAD",
    EDUCATION = "EDUCATION",
    CHATGPT = "CHATGPT",
    GEMINI = "GEMINI",
    PRESENTATIONCONTENT = "PRESENTATIONCONTENT",
    LINKEDINPLATFORM = "LINKEDINPLATFORM",
    SESSIONTOKENS = "SESSIONTOKENS",
    POSITIONSLIKED = "POSITIONSLIKED",
    BLOGSFAVORITE = "BLOGSFAVORITE",
    BLOGPOSTCOMMENT = "BLOGPOSTCOMMENT",
    CUSTOMERBRIEF = "CUSTOMERBRIEF",
    BRIEFTEMPLATE = "BLOGPOSTCOMMENT",
    BRIEFQUESTION = "BRIEFQUESTION",
    BRIEFSECTION = "BRIEFSECTION",
    USERCREDENTIALS = "USERCREDENTIALS",
    INDUSTRYDOMAINS = "INDUSTRYDOMAINS",
    CURRENTCOMPANY = "CURRENTCOMPANY",
    COMPANYVERSIONS = "COMPANYVERSIONS",
    COMPANYBENEFITS = "COMPANYBENEFITS",
    COMPANYVALUES = "COMPANYVALUES",
    POSITIONPIPELINE = "POSITIONPIPELINE",
    PIPELINESTAGE = "PIPELINESTAGE",
    PIPELINESTAGEFEEDBACK = "PIPELINESTAGEFEEDBACK",
    POSITIONWORKFLOW = "POSITIONWORKFLOW",
    POSITIONWORKFLOWSTAGE = "POSITIONWORKFLOWSTAGE",
    TALENTPIPELINEPROGRESS = "TALENTPIPELINEPROGRESS",
    CHATROOMS = "CHATROOMS",
    CHATMESSAGES = "CHATMESSAGES",
    VIDEOCHATROOMS = "VIDEOCHATROOMS",
    VIDEOCHATROOMPARTICIPANTS = "VIDEOCHATROOMPARTICIPANTS",
    VIDEORECORDS = "VIDEORECORDS",
    TELEGRAM = "TELEGRAM",
    TELEGRAMNOTIFICATIONS = "TELEGRAMNOTIFICATIONS",
    WHATSAPP = "WHATSAPP",
    SCREENINGFORMS = "SCREENINGFORMS",
    SCREENINGQUESTIONS = "SCREENINGQUESTIONS",
    SCREENINGQUESTIONTEMPLATES = "SCREENINGQUESTIONTEMPLATES",
    SCREENINGRESPONSES = "SCREENINGRESPONSES",
    TALENTNOTES = "TALENTNOTES",
    HEALTHCHECK = "HEALTHCHECK",
    PAYMENTS = "PAYMENTS",
    LIVECODINGSESSIONS = "LIVECODINGSESSIONS",
    CODESNIPPETS = "CODESNIPPETS",
    RECRUITERACTIVITYACCESS = "RECRUITERACTIVITYACCESS"
}