// tab-frontend/src/app/modules/positions/models/layout-type.ts
export enum LayoutType {
    fullView = "Full View", 
    shortView = "Short View",
    tileView = "Tile View", 
    tableView = "Table View"
}

export enum LayoutTypeIcons {
    fullView = "zoom_out_map", // pageview zoom_out_map
    shortView = "short_text", // segment
    tileView = "chrome_reader_mode", // visibility article description  
    tableView = "border_all" // visibility table_chart
}